requirements.txt
requests==2.23.0
peewee==3.13.1
mysqlclient==1.4.6
apscheduler==3.6.3
flask==1.1.1
prometheus-flask-exporter==0.13.0
swagger-ui-py==0.2.1